requirements.txt
asgiref==3.2.3
attrs==19.3.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
descartes==1.1.0
dj-database-url==0.5.0
Django==3.0.1
django-heroku==0.3.1
django-rest==0.0.1
djangorestframework==3.11.0
expects==0.9.0
Fiona==1.8.13
geopandas==0.6.2
gunicorn==20.0.4
importlib-metadata==1.3.0
kiwisolver==1.1.0
matplotlib==3.1.2
more-itertools==8.0.2
munch==2.5.0
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pyproj==2.4.2.post1
pytest==5.3.2
pytest-dotenv==0.4.0
pytest-env==0.6.2
pytest-expect==1.1.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
Shapely==1.6.4.post2
six==1.13.0
sqlparse==0.3.0
u-msgpack-python==2.5.2
wcwidth==0.1.7
whitenoise==5.0.1
zipp==0.6.0
